requirements.txt
# pip install flask-login flask-bcrypt
# pip install flask flask-login flask-sqlalchemy google-generativeai pillow python-dotenv
flask
flask-login
flask-bcrypt
flask-sqlalchemy
google-generativeai
pillow 
python-dotenv